+++ joueur_medium.hh
@@ -14,7 +14,7 @@ class Joueur_medium :public Joueur_debutant{
 		int suspicion();
 		int score();
 	private:
-		int _taux_suspicion;
+		float _taux_suspicion;
 		int _score;
 
 };
