+++ transaction_fake.hh
@@ -10,5 +10,5 @@ public:
     int montant();
     int suspicion();
 protected:
-    int _suspicionRate;
+    float _suspicionRate;
 };
